Remove cache key module and store permid cache key in buffer
